refactor: album user repository
