+++ server/src/types.ts
@@ -2,6 +2,7 @@ import { UserEntity } from 'src/entities/user.entity';
 import { ExifOrientation, ImageFormat, Permission, TranscodeTarget, VideoCodec } from 'src/enum';
 import { AccessRepository } from 'src/repositories/access.repository';
 import { ActivityRepository } from 'src/repositories/activity.repository';
+import { AlbumUserRepository } from 'src/repositories/album-user.repository';
 import { ApiKeyRepository } from 'src/repositories/api-key.repository';
 import { AuditRepository } from 'src/repositories/audit.repository';
 import { ConfigRepository } from 'src/repositories/config.repository';
@@ -31,6 +32,7 @@ export type RepositoryInterface<T extends object> = Pick<T, keyof T>;
 
 export type IActivityRepository = RepositoryInterface<ActivityRepository>;
 export type IAccessRepository = { [K in keyof AccessRepository]: RepositoryInterface<AccessRepository[K]> };
+export type IAlbumUserRepository = RepositoryInterface<AlbumUserRepository>;
 export type IApiKeyRepository = RepositoryInterface<ApiKeyRepository>;
 export type IAuditRepository = RepositoryInterface<AuditRepository>;
 export type IConfigRepository = RepositoryInterface<ConfigRepository>;
